requirements2.txt
absl-py==0.10.0
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
cycler==0.10.0
Cython==0.29.21
gluoncv==0.6.0
google-auth==1.22.0
google-auth-oauthlib==0.4.1
grpcio==1.31.0
idna==2.6
importlib-metadata==2.0.0
joblib==1.0.0
kiwisolver==1.2.0
Markdown==3.3.2
matplotlib==3.3.2
mkl-service==2.3.0
multidict==4.7.5
mxnet-cu100==1.5.0
numpy==1.19.5
oauthlib==3.0.1
olefile==0.46
pandas==1.2.1
patsy==0.5.1
Pillow==8.0.0
pip==20.2.4
portalocker==1.5.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.4.2.post1
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.1
scipy==1.5.2
seaborn==0.11.1
setuptools==49.6.0.post20201009
six==1.15.0
statsmodels==0.12.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.50.2
typing-extensions==3.7.4.3
urllib3==1.22
Werkzeug==1.0.1
wheel==0.35.1
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.5.1
zipp==3.3.1